feat: 요구사항에 맞게 로비에서 종료된 게임에 삭제 요청시 비밀번호 입력하고 삭제할 수 있도록 변경

+++ src/main/java/chess/controller/spring/GameController.java
@@ -120,6 +120,10 @@ public class GameController {
     @DeleteMapping("/{gameId}")
     @ResponseBody
     public ResponseEntity<Long> deleteGame(@PathVariable Long gameId) {
+        final ChessGame game = gameService.findByGameId(gameId);
+        if (!game.isEnd()) {
+            throw new IllegalStateException("게임이 종료되지 않아서 삭제할 수 없습니다.");
+        }
         final Long deletedId = gameService.deleteGame(gameId);
         return ResponseEntity.ok(deletedId);
     }
